+++ client/web/src/enterprise/batches/list/BatchChangesListIntro.tsx
@@ -6,6 +6,8 @@ import { CardBody, Card, Link } from '@sourcegraph/wildcard'
 
 import { SourcegraphIcon } from '../../../auth/icons'
 
+import { BatchChangesChangelogAlert } from './BatchChangesChangelogAlert'
+
 import styles from './BatchChangesListIntro.module.scss'
 
 export interface BatchChangesListIntroProps {
@@ -13,15 +15,26 @@ export interface BatchChangesListIntroProps {
 }
 
 export const BatchChangesListIntro: React.FunctionComponent<BatchChangesListIntroProps> = ({ isLicensed }) => {
-    if (isLicensed === undefined || isLicensed === true) {
+    if (isLicensed === undefined) {
         return null
     }
 
     return (
         <div className="row">
-            <div className="col-12 col-md-6 mb-3">
-                <BatchChangesUnlicensedAlert />
-            </div>
+            {isLicensed === true ? (
+                <div className="col-12">
+                    <BatchChangesChangelogAlert />
+                </div>
+            ) : (
+                <>
+                    <div className="col-12 col-md-6 mb-3">
+                        <BatchChangesUnlicensedAlert />
+                    </div>
+                    <div className="col-12 col-md-6 mb-3">
+                        <BatchChangesChangelogAlert />
+                    </div>
+                </>
+            )}
         </div>
     )
 }