+++ client/web/src/enterprise/insights/components/insights-view-grid/components/backend-insight/components/index.ts
@@ -1,0 +1,7 @@
+export type { DrillDownInsightCreationFormValues } from './drill-down-filters-panel/components/drill-down-insight-creation-form/DrillDownInsightCreationForm'
+
+export { DrillDownFiltersPanel } from './drill-down-filters-panel/DrillDownFiltersPanel'
+export { DrillDownFiltersPopover } from './drill-down-filters-popover/DrillDownFiltersPopover'
+
+export { BackendInsightErrorAlert, BackendAlertOverlay } from './backend-insight-alerts/BackendInsightAlerts'
+export { BackendInsightChart } from './backend-insight-chart/BackendInsightChart'