+++ client/web/src/enterprise/insights/pages/insights/creation/search-insight/utils/insight-sanitizer.ts
@@ -39,6 +39,7 @@ export function getSanitizedSearchInsight(rawInsight: CreateInsightFormFields): 
             filters: {
                 excludeRepoRegexp: '',
                 includeRepoRegexp: '',
+                contexts: [],
                 repositories: [],
             },
         }