+++ dev/sg/main.go
@@ -10,50 +10,49 @@ import (
 	"github.com/urfave/cli/v2"
 
 	"github.com/sourcegraph/sourcegraph/dev/sg/internal/secrets"
+	"github.com/sourcegraph/sourcegraph/dev/sg/internal/sgconf"
 	"github.com/sourcegraph/sourcegraph/dev/sg/internal/stdout"
 	"github.com/sourcegraph/sourcegraph/dev/sg/root"
-	"github.com/sourcegraph/sourcegraph/lib/output"
+	"github.com/sourcegraph/sourcegraph/lib/errors"
 )
 
 func main() {
-	if err := loadSecrets(); err != nil {
-		writeWarningLinef("failed to open secrets: %s", err)
+	if os.Args[len(os.Args)-1] == "--generate-bash-completion" {
+		batchCompletionMode = true
 	}
-	ctx := secrets.WithContext(context.Background(), secretsStore)
 
-	if err := sg.RunContext(ctx, os.Args); err != nil {
+	if err := sg.RunContext(context.Background(), os.Args); err != nil {
 		fmt.Printf("error: %s\n", err)
 		os.Exit(1)
 	}
 }
 
-const (
-	defaultConfigFile          = "sg.config.yaml"
-	defaultConfigOverwriteFile = "sg.config.overwrite.yaml"
-	defaultSecretsFile         = "sg.secrets.json"
-)
-
 var (
 	BuildCommit = "dev"
 
-	// globalConf is the global config. If a command needs to access it, it *must* call
-	// `parseConf` before.
-	globalConf *Config
-
-	// secretsStore is instantiated when sg gets run.
-	secretsStore *secrets.Store
-
-	// Note that these values are only available after the main sg CLI app has been run.
-	configFlag          string
-	overwriteConfigFlag string
-	skipAutoUpdatesFlag bool
+	// configFile is the path to use with sgconf.Get - it must not be used before flag
+	// initialization.
+	configFile string
+	// configOverwriteFile is the path to use with sgconf.Get - it must not be used before
+	// flag initialization.
+	configOverwriteFile string
 
 	// Global verbose mode
 	verbose bool
 
 	// postInitHooks is useful for doing anything that requires flags to be set beforehand,
-	// e.g. generating help text based on parsed config
-	postInitHooks []cli.ActionFunc
+	// e.g. generating help text based on parsed config, and are called before any command
+	// Action is executed. These should run quickly and must fail gracefully.
+	//
+	// Commands can register postInitHooks in an 'init()' function that appends to this
+	// slice.
+	postInitHooks []func(cmd *cli.Context)
+
+	// batchCompletionMode determines if we are in bash completion mode. In this mode,
+	// sg should respond quickly, so most setup tasks (e.g. postInitHooks) are skipped.
+	//
+	// Do not run complicated tasks, etc. in Before or After hooks when in this mode.
+	batchCompletionMode bool
 )
 
 // sg is the main sg CLI application.
@@ -73,49 +72,72 @@ var sg = &cli.App{
 		},
 		&cli.StringFlag{
 			Name:        "config",
+			Aliases:     []string{"c"},
 			Usage:       "load sg configuration from `file`",
 			EnvVars:     []string{"SG_CONFIG"},
 			TakesFile:   true,
-			Value:       defaultConfigFile,
-			Destination: &configFlag,
+			Value:       sgconf.DefaultFile,
+			Destination: &configFile,
 		},
 		&cli.StringFlag{
 			Name:        "overwrite",
+			Aliases:     []string{"o"},
 			Usage:       "load sg configuration from `file` that is gitignored and can be used to, for example, add credentials",
 			EnvVars:     []string{"SG_OVERWRITE"},
 			TakesFile:   true,
-			Value:       defaultConfigOverwriteFile,
-			Destination: &overwriteConfigFlag,
+			Value:       sgconf.DefaultOverwriteFile,
+			Destination: &configOverwriteFile,
 		},
 		&cli.BoolFlag{
-			Name:        "skip-auto-update",
-			Usage:       "prevent sg from automatically updating itself",
-			EnvVars:     []string{"SG_SKIP_AUTO_UPDATE"},
-			Value:       BuildCommit == "dev", // Default to skip in dev, otherwise don't
-			Destination: &skipAutoUpdatesFlag,
+			Name:    "skip-auto-update",
+			Usage:   "prevent sg from automatically updating itself",
+			EnvVars: []string{"SG_SKIP_AUTO_UPDATE"},
+			Value:   BuildCommit == "dev", // Default to skip in dev, otherwise don't
 		},
 	},
 	Before: func(cmd *cli.Context) error {
+		if batchCompletionMode {
+			// All other setup pertains to running commands - to keep completions fast,
+			// we skip all other setup.
+			return nil
+		}
+
 		if verbose {
 			stdout.Out.SetVerbose()
 		}
 
-		// We always try to set this, since we
-		// often want to watch files, start commands, etc...
+		// Validate configuration flags, which is required for sgconf.Get to work everywhere else.
+		if configFile == "" {
+			return errors.Newf("--config must not be empty")
+		}
+		if configOverwriteFile == "" {
+			return errors.Newf("--overwrite must not be empty")
+		}
+
+		// Set up access to secrets
+		secretsStore, err := loadSecrets()
+		if err != nil {
+			writeWarningLinef("failed to open secrets: %s", err)
+		} else {
+			cmd.Context = secrets.WithContext(cmd.Context, secretsStore)
+		}
+
+		// We always try to set this, since we often want to watch files, start commands, etc...
 		if err := setMaxOpenFiles(); err != nil {
 			writeWarningLinef("Failed to set max open files: %s", err)
 		}
 
+		// Check for updates, unless we are running update manually.
 		if cmd.Args().First() != "update" {
-			// If we're not running "sg update ...", we want to check the version first
-			err := checkSgVersion(cmd.Context)
+			err := checkSgVersionAndUpdate(cmd.Context, cmd.Bool("skip-auto-update"))
 			if err != nil {
-				writeWarningLinef("Checking sg version and updating failed: %s", err)
+				writeWarningLinef("update check: %s", err)
 				// Do not exit here, so we don't break user flow when they want to
 				// run `sg` but updating fails
 			}
 		}
 
+		// Call registered hooks last
 		for _, hook := range postInitHooks {
 			hook(cmd)
 		}
@@ -153,6 +175,8 @@ var sg = &cli.App{
 		funkyLogoCommand,
 	},
 
+	CommandNotFound: suggestCommands,
+
 	EnableBashCompletion:   true,
 	UseShortOptionHandling: true,
 
@@ -160,51 +184,11 @@ var sg = &cli.App{
 	HideHelpCommand: true,
 }
 
-func loadSecrets() error {
+func loadSecrets() (*secrets.Store, error) {
 	homePath, err := root.GetSGHomePath()
 	if err != nil {
-		return err
+		return nil, err
 	}
-	fp := filepath.Join(homePath, defaultSecretsFile)
-	secretsStore, err = secrets.LoadFile(fp)
-	return err
-}
-
-// parseConf parses the config file and the optional overwrite file.
-// Iear the conf has already been parsed it's a noop.
-func parseConf(confFile, overwriteFile string) (bool, output.FancyLine) {
-	if globalConf != nil {
-		return true, output.FancyLine{}
-	}
-
-	// Try to determine root of repository, so we can look for config there
-	repoRoot, err := root.RepositoryRoot()
-	if err != nil {
-		return false, output.Linef("", output.StyleWarning, "Failed to determine repository root location: %s", err)
-	}
-
-	// If the configFlag/overwriteConfigFlag flags have their default value, we
-	// take the value as relative to the root of the repository.
-	if confFile == defaultConfigFile {
-		confFile = filepath.Join(repoRoot, confFile)
-	}
-
-	if overwriteFile == defaultConfigOverwriteFile {
-		overwriteFile = filepath.Join(repoRoot, overwriteFile)
-	}
-
-	globalConf, err = ParseConfigFile(confFile)
-	if err != nil {
-		return false, output.Linef("", output.StyleWarning, "Failed to parse %s%s%s%s as configuration file:%s\n%s", output.StyleBold, confFile, output.StyleReset, output.StyleWarning, output.StyleReset, err)
-	}
-
-	if ok, _ := fileExists(overwriteFile); ok {
-		overwriteConf, err := ParseConfigFile(overwriteFile)
-		if err != nil {
-			return false, output.Linef("", output.StyleWarning, "Failed to parse %s%s%s%s as overwrites configuration file:%s\n%s", output.StyleBold, overwriteFile, output.StyleReset, output.StyleWarning, output.StyleReset, err)
-		}
-		globalConf.Merge(overwriteConf)
-	}
-
-	return true, output.FancyLine{}
+	fp := filepath.Join(homePath, secrets.DefaultFile)
+	return secrets.LoadFromFile(fp)
 }