+++ dev/sg/os.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"os"
 	"syscall"
 
 	"github.com/sourcegraph/sourcegraph/lib/errors"
@@ -28,15 +27,4 @@ func setMaxOpenFiles() error {
 	}
 
 	return nil
-}
-
-func fileExists(path string) (bool, error) {
-	_, err := os.Stat(path)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return false, nil
-		}
-		return false, err
-	}
-	return true, nil
 }