+++ dev/sg/sg_live.go
@@ -20,6 +20,9 @@ var liveCommand = &cli.Command{
 	Category:    CategoryCompany,
 	Description: constructLiveCmdLongHelp(),
 	Action:      execAdapter(liveExec),
+	BashComplete: completeOptions(func() (options []string) {
+		return append(environmentNames(), `https\://...`)
+	}),
 }
 
 func constructLiveCmdLongHelp() string {
@@ -28,7 +31,7 @@ func constructLiveCmdLongHelp() string {
 	fmt.Fprintf(&out, "Prints the Sourcegraph version deployed to the given environment.")
 	fmt.Fprintf(&out, "\n")
 	fmt.Fprintf(&out, "\n")
-	fmt.Fprintf(&out, "AVAILABLE PRESET ENVIRONMENTS\n")
+	fmt.Fprintf(&out, "AVAILABLE PRESET ENVIRONMENTS:\n")
 
 	for _, name := range environmentNames() {
 		fmt.Fprintf(&out, "  %s\n", name)