+++ dev/sg/sg_version.go
@@ -13,6 +13,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/dev/sg/internal/run"
 	"github.com/sourcegraph/sourcegraph/dev/sg/internal/stdout"
 	"github.com/sourcegraph/sourcegraph/dev/sg/root"
+	"github.com/sourcegraph/sourcegraph/lib/errors"
 	"github.com/sourcegraph/sourcegraph/lib/output"
 )
 
@@ -28,7 +29,7 @@ var (
 		Subcommands: []*cli.Command{
 			{
 				Name:    "changelog",
-				Aliases: []string{"changes"},
+				Aliases: []string{"c"},
 				Usage:   "See what's changed in or since this version of sg",
 				Flags: []cli.Flag{
 					&cli.BoolFlag{
@@ -39,6 +40,7 @@ var (
 					&cli.IntFlag{
 						Name:        "limit",
 						Usage:       "Number of changelog entries to show.",
+						Value:       5,
 						Destination: &versionChangelogEntries,
 					},
 				},
@@ -54,6 +56,10 @@ func versionExec(ctx context.Context, args []string) error {
 }
 
 func changelogExec(ctx context.Context, args []string) error {
+	if _, err := run.GitCmd("fetch", "origin", "main"); err != nil {
+		return errors.Newf("failed to update main: %s", err)
+	}
+
 	logArgs := []string{
 		// Format nicely
 		"log", "--pretty=%C(reset)%s %C(dim)%h by %an, %ar",
@@ -98,9 +104,13 @@ func changelogExec(ctx context.Context, args []string) error {
 	return nil
 }
 
-const sgOneLineCmd = `curl --proto '=https' --tlsv1.2 -sSLf https://install.sg.dev | sh`
+func checkSgVersionAndUpdate(ctx context.Context, skipUpdate bool) error {
+	if BuildCommit == "dev" {
+		// If `sg` was built with a dirty `./dev/sg` directory it's a dev build
+		// and we don't need to display this message.
+		return nil
+	}
 
-func checkSgVersion(ctx context.Context) error {
 	_, err := root.RepositoryRoot()
 	if err != nil {
 		// Ignore the error, because we only want to check the version if we're
@@ -108,18 +118,15 @@ func checkSgVersion(ctx context.Context) error {
 		return nil
 	}
 
-	if BuildCommit == "dev" {
-		// If `sg` was built with a dirty `./dev/sg` directory it's a dev build
-		// and we don't need to display this message.
-		return nil
-	}
-
 	rev := strings.TrimPrefix(BuildCommit, "dev-")
-	out, err := run.GitCmd("rev-list", fmt.Sprintf("%s..origin/main", rev), "./dev/sg")
+	out, err := run.GitCmd("rev-list", fmt.Sprintf("%s..origin/main", rev), "--", "./dev/sg")
 	if err != nil {
-		fmt.Printf("error getting new commits since %s in ./dev/sg: %s\n", rev, err)
-		fmt.Printf("try reinstalling sg with `%s`.\n", sgOneLineCmd)
-		os.Exit(1)
+		if strings.Contains(out, "bad revision") {
+			// installed revision is not available locally, that is fine - we wait for the
+			// user to eventually do a fetch
+			return errors.New("current sg version not found - you may want to run 'git fetch origin main'.")
+		}
+		return err
 	}
 
 	out = strings.TrimSpace(out)
@@ -128,7 +135,7 @@ func checkSgVersion(ctx context.Context) error {
 		return nil
 	}
 
-	if skipAutoUpdatesFlag {
+	if skipUpdate {
 		stdout.Out.WriteLine(output.Linef("", output.StyleSearchMatch, "╭──────────────────────────────────────────────────────────────────╮  "))
 		stdout.Out.WriteLine(output.Linef("", output.StyleSearchMatch, "│                                                                  │░░"))
 		stdout.Out.WriteLine(output.Linef("", output.StyleSearchMatch, "│ HEY! New version of sg available. Run 'sg update' to install it. │░░"))
@@ -142,7 +149,11 @@ func checkSgVersion(ctx context.Context) error {
 	stdout.Out.WriteLine(output.Line(output.EmojiInfo, output.StyleSuggestion, "Auto updating sg ..."))
 	newPath, err := updateToPrebuiltSG(ctx)
 	if err != nil {
-		return err
+		return errors.Newf("failed to install update: %s", err)
 	}
+	writeSuccessLinef("sg has been updated!")
+	stdout.Out.Write("To see what's new, run 'sg version changelog'.")
+
+	// Run command with new binary
 	return syscall.Exec(newPath, os.Args, os.Environ())
 }